count the first linear step in stdmultilevelsolver
To be able to reliably set relative tolerances for the linear solver
we have to evaluate the initial residual _before_ doing any
mg-iterations. Otherwise linear_tol might never be reached, for
example if the umfpack direct solver is used.

With this change the default linear_tol can be set to the precision of
matrixentrytype (float or double) and linear_globaltol can be set to a
very small value, such as 1e-40.

This change results in one more linear residual evaluation per linear
solve, which should be negligible.