finally switch from deprecated hash_map to new unordered_map
This completes the switch started by winni for NEWER_THAN_GCC_4_2.

First an important warning:

! Vertex numbers in refined grids will change (with NEWER_THAN_GCC_4_2) !

If you have code that reads vectors from bup files and use a grid generated
from an inp file and prerefine you will get problems.
From now on, if you define NEWER_THAN_GCC_4_2 in your cmake config, you will
use the new TR1 unordered_map instead of the deprecated hash_map everywhere.
This gets rid of the warnings with recent gcc versions. However, assignment of new
vertex numbers (in the method update_news) depends on the order in which the iterator
traverses the hangmap, which is implementation defined.
You should always save and load the gup corresponding to the bup!

This change also uncovered a very strange bug that is fixed in the rest of the commit:
In hierachicalmesh{2,3}.cc the method {Line,Quad}Hang.find(i) was called for an
integer i. This made no sense since the keys of {Line,Quad}Hang are
Edges or Faces with N = 2,4 elements. However, the corresponding class EdgeArray<N>
had a superfluous constructor for an integer argument so that we ended up looking
for an edge of the form {i,i}, which obviously does not exist in LineHang. So since
{i,i} != {j,k} for any real edge {j,k} with j \neq k, this assertion (and
corresponding if clause) was guaranteed to be false up to now.

Now, for some reason operator== in EdgeArray is implemented in such a way that
{i,j} == {i,i} evaluates to false
_but_
{i,i} == {i,j} evaluates to true.
The old hash_map internally uses the former comparison in its implementation of find(),
the new unordered unordered_map the latter, so this assertion started to trigger.

Since the find was guaranteed to fail up to now and the intent of the original author
is unclear, I just removed the problematic find() statements.
To prevent such mixup in the future the superfluous constructors were removed
to forbid creation of invalid edges.
Also some old or superfluous code fragments were cleaned up.